requirements.txt
asgiref==3.10.0
Django==5.2.7
django-rest-framework==0.1.0
djangorestframework==3.16.1
dotenv==0.9.9
python-dotenv==1.2.1
sqlparse==0.5.3
